Add files via upload
Initial tests with a genetic approach minimizing the singlet probability by tuning the applied external magnetic field pulses with respect to time by using a genetic approach. This genetic approach is also fine-tuned, similar to a hypervalidation setup, in which different configurations are tested and the best is selected according to a given measure (the singlet probability) by using a grid-search approach.